Added Pepsysco data: 2022 10.1021/acsomega.2c02425
This raw data very kindly provided by the authors to Rebecca Birolo.

Currently, the model does significantly reduce the loss in making the
prediction, but doesn't really appear to start predicting the exact
products. I think this is because we need to do more work on the
Monte-Carlo synthesis part of the code, and think a bit harder about
what sort of byproducts you get in the crude product MassSpec.
Pepsysco does direct mass-spec on the crude (electrospray?) so the yield
is likely biased towards what is ionisable rather than what was synth'd
(Anna Barnard).
But clean and dry, so anion from TFA won't be there, but perhaps could
often see formate (so +44... but does it substitute?)?
Certainly the first entry in the data shows a +38 Da peak above the pure
product... what is this?